+++ src/Sulu/Bundle/MediaBundle/Resources/public/js/components/media-selection/main.js
@@ -69,7 +69,9 @@ define(['underscore'], function(_) {
 
                 if (thumbnail) {
                     content.push([
-                        '    <img src="', thumbnail, '"/>'
+                        '<span class="image">',
+                        '    <img src="', thumbnail, '"/>',
+                        '</span>'
                     ].join(''));
                 }